% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_archive_function.R
\name{daily_archive}
\alias{daily_archive}
\title{This function retrieves hourly data from aswe sites, including both archived and current year data}
\usage{
daily_archive(
  parameter = c("swe", "snow_depth", "precipitation", "temperature"),
  get_year = "All",
  id
)
}
\arguments{
\item{parameter}{Defines the parameter (type of data) your want to retrieve}

\item{get_year}{Define the year that you want to retrieve. Defaults to "All"}

\item{id}{Station ID you are looking for}
}
\description{
This function retrieves hourly data from aswe sites, including both archived and current year data
}
\examples{
\dontrun{}
}
\keyword{internal}
