% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_current_function.R
\name{daily_current}
\alias{daily_current}
\title{This function retrieves daily data from aswe sites for the current water year only.
Daily values are calculated as the daily mean except for temperature, where the daily max and min are retrieved and for swe, where the 16:00 UTC measurement is used}
\usage{
daily_current(
  parameter = c("swe", "snow_depth", "precipitation", "temperature"),
  id
)
}
\arguments{
\item{parameter}{Defines the parameter (type of data) your want to retrieve}

\item{id}{Station ID you are looking for}
}
\description{
This function retrieves daily data from aswe sites for the current water year only.
Daily values are calculated as the daily mean except for temperature, where the daily max and min are retrieved and for swe, where the 16:00 UTC measurement is used
}
\examples{
\dontrun{}
}
\keyword{internal}
