% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manual_databc_function.R
\name{get_manual_swe}
\alias{get_manual_swe}
\title{Get BC Manual Snow Data}
\usage{
get_manual_swe(
  station_id,
  survey_period = "All",
  get_year = "All",
  use_archive = c("Yes", "No"),
  update_archive = c("Yes", "No"),
  directory_archive = "./data/cache/"
)
}
\arguments{
\item{station_id}{Define the station id you want. Can be an individual site, a string of site IDs, or all ASWE sites. Defaults to "All".}

\item{survey_period}{Survey period you want to retrieve. Defaults to 'All'}

\item{get_year}{Water year of data you want to retrieve. Defaults to 'All'}

\item{use_archive}{Whether you want to cache data to speed downstream calculations. Defaults to c("Yes", "No"),}

\item{update_archive}{Whether you want to update the historical archive (data before this water year). Can choose either yes or no. If the archive is greater than 30 days old it will automatically update}

\item{directory_archive}{The directory where the data cache will be saved. Defaults to "./data/cache/" of your active directory}
}
\description{
This function allows you to obtain data for . It retrieves this data from Data BC
}
\examples{
get_manual_swe()
}
\keyword{DataBC}
\keyword{Manual}
\keyword{data}
\keyword{snow}
