% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snow_function.R
\name{get_snow}
\alias{get_snow}
\title{Get Snow Function
Function that determines whether the user has specified that they want data for a manual or ASWE site}
\usage{
get_snow(
  id = c("All", "automated", "manual"),
  get_year = "All",
  survey_period = "All",
  parameter = c("swe", "snow_depth", "precipitation", "temperature"),
  timestep = c("hourly", "daily")
)
}
\arguments{
\item{id}{Define the station id you want. Can be an individual site, a string of site IDs, or all ASWE sites. Defaults to "All"; this will return a great deal of data.}

\item{get_year}{Define the year that you want to retrieve. Defaults to "All"}

\item{survey_period}{The manual survey period a user wants. Defaults to "all"}

\item{parameter}{Defaults to: "swe", "snow_depth", "precipitation", "temperature". Type of data you want to retrieve}

\item{timestep}{Whether the user wants the hourly or daily data. Choices are "hourly" or "daily"}
}
\description{
Get Snow Function
Function that determines whether the user has specified that they want data for a manual or ASWE site
}
\examples{
\dontrun{}
}
\keyword{Get}
\keyword{data}
\keyword{snow}
