% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_current_function.R
\name{hourly_current}
\alias{hourly_current}
\title{This function retrieves hourly data from aswe sites for the current water year only}
\usage{
hourly_current(
  parameter = c("swe", "snow_depth", "precipitation", "temperature"),
  id
)
}
\arguments{
\item{parameter}{Defines the parameter (type of data) your want to retrieve}

\item{id}{Station ID you are looking for}
}
\description{
This function retrieves hourly data from aswe sites for the current water year only
}
\examples{
\dontrun{}
}
\keyword{internal}
