% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_basinlocations_function.R
\name{snow_basin_areas}
\alias{snow_basin_areas}
\title{Get the location of snow basins (for mapping)}
\usage{
snow_basin_areas(crs = "3005")
}
\arguments{
\item{crs}{Defaults to 3005}
}
\description{
This function gets the location of snow surveys from the Data BC catalogue
}
\examples{
snow_basin_areas()
}
\keyword{Get}
\keyword{locations}
\keyword{snow}
