% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manual_data_old.R
\name{snow_manual_archive}
\alias{snow_manual_archive}
\title{Get manual snow data from archive - old version}
\usage{
snow_manual_archive()
}
\description{
Get manual snow data from archive - old version
}
\examples{
\dontrun{}
}
\keyword{Get}
\keyword{archive}
\keyword{manual}
