% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manual_data_old.R
\name{snow_manual_current}
\alias{snow_manual_current}
\title{Get manual snow data from current year file - old version}
\usage{
snow_manual_current()
}
\description{
Get manual snow data from current year file - old version
}
\examples{
\dontrun{}
}
\keyword{Get}
\keyword{current}
\keyword{manual}
