% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_function_wateryear.R
\name{wtr_yr}
\alias{wtr_yr}
\title{Assign water year function
This function assigns the water year according to a column or vector of dates}
\usage{
wtr_yr(dates, start_month = 10)
}
\arguments{
\item{dates}{A column of dates (in date format)}

\item{start_month}{The month that the water year starts on. Defaults to 10 (October)}
}
\description{
Assign water year function
This function assigns the water year according to a column or vector of dates
}
\examples{
wtr_yr()
}
