% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aswe_databc_function.R
\name{get_aswe_databc}
\alias{get_aswe_databc}
\title{Get BC ASWE Data Function. This function allows you to obtain data from BC ASWE sites. It retrieves this data from Data BC}
\usage{
get_aswe_databc(
  station_id = "All",
  get_year = "All",
  parameter = c("swe", "snow_depth", "precipitation", "temperature"),
  timestep = c("hourly", "daily")
)
}
\arguments{
\item{station_id}{Define the station id you want. Can be an individual site, a string of site IDs, or all ASWE sites. Defaults to "All"; this will return a great deal of data.}

\item{get_year}{Define the year that you want to retrieve. Defaults to "All"}

\item{parameter}{Defaults to: "swe", "snow_depth", "precipitation", "temperature". Type of data you want to retrieve}

\item{timestep}{Whether the user wants the hourly or daily data. Choices are "hourly" or "daily"}
}
\description{
Get BC ASWE Data Function. This function allows you to obtain data from BC ASWE sites. It retrieves this data from Data BC
}
\examples{
\dontrun{}
}
\keyword{ASWE}
\keyword{Data}
\keyword{Get}
