% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aswe_databc_function.R
\name{get_aswe_databc}
\alias{get_aswe_databc}
\title{Get BC ASWE Data Function}
\usage{
get_aswe_databc(
  station_id = "All",
  get_year = "All",
  parameter_id = c("SWE", "Snow_Depth", "Precipitation", "Temperature"),
  force = FALSE,
  ask = TRUE,
  ...
)
}
\arguments{
\item{station_id}{Define the station id you want. Can be an individual site, a string of site IDs, or all ASWE sites. Defaults to "All"; this will return a great deal of data.}

\item{get_year}{Define the year that you want to retrieve. Defaults to "All"}

\item{parameter_id}{Defaults to: "SWE", "Snow_Depth", "Precipitation", "Temperature". Type of data you want to retrieve}

\item{force}{Whether you want to re-download the archived file whether it is updated or no. Defaults to FALSE}

\item{ask}{Whether the user is asked whether to create a directory for the cached file. Defaults to TRUE}
}
\description{
This function allows you to obtain data from BC ASWE sites. It retrieves this data from Data BC
}
\examples{
get_aswe_databc()
}
\keyword{ASWE}
\keyword{Data}
\keyword{Get}
