% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manual_databc_function.R
\name{get_manual_swe}
\alias{get_manual_swe}
\title{Get BC Manual Snow Data}
\usage{
get_manual_swe(station_id = "All", survey_period = "All", get_year = "All")
}
\arguments{
\item{station_id}{Define the station id you want. Can be an individual site, a string of site IDs ("4A10"), or all ASWE sites ("All"). Defaults to "All".}

\item{survey_period}{Survey period you want to retrieve. Can choose either all sites ("All") or a specific survey period ("01-Jan", "01-Feb", "01-Mar", "01-Apr", "01-May", "15-May", "01-Jun", "15-Jun"). Defaults to 'All'}

\item{get_year}{Water year of data you want to retrieve. Defaults to 'All'}
}
\description{
This function allows you to obtain data for . It retrieves this data from Data BC
}
\examples{
\dontrun{}
get_manual_swe()
}
\keyword{DataBC}
\keyword{Manual}
\keyword{data}
\keyword{snow}
