% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_autolocation_function.R
\name{snow_auto_location}
\alias{snow_auto_location}
\title{This function gets the location of snow surveys from the Data BC catalog}
\usage{
snow_auto_location()
}
\description{
This function gets the location of snow surveys from the Data BC catalog
}
\examples{
\dontrun{}
snow_auto_location()
}
\keyword{Get}
\keyword{automated}
\keyword{locations}
\keyword{snow}
\keyword{survey}
