% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_manuallocation_function.R
\name{snow_manual_location}
\alias{snow_manual_location}
\title{This function gets the location of snow surveys from the Data BC catalog}
\usage{
snow_manual_location()
}
\description{
This function gets the location of snow surveys from the Data BC catalog
}
\examples{
\dontrun{}
snow_manual_location()
}
\keyword{Get}
\keyword{locations}
\keyword{manual}
\keyword{snow}
\keyword{survey}
